pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ishtar-ai"
version = "0.1.0"
description = "Ishtar AI: RAG + Multi-Agent API (bootstrap)"
authors = [{name="David Stroud"}]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "httpx>=0.27",
  "tenacity>=8.3",
  "numpy>=1.26",
  "faiss-cpu>=1.8.0",
  "langchain>=0.2.14",
  "langgraph>=0.2.35",
  "openai>=1.43.0",
  "prometheus-client>=0.20",
  "python-dotenv>=1.0",
  "tiktoken>=0.7",
  "feedparser>=6.0",
]
[tool.setuptools.packages.find]
where = ["."]
